prepare to define Dimension in raxiom
