+++ diman/Cargo.toml
@@ -16,3 +16,4 @@ hdf5 = {version = "0.8.1", optional = true }
 mpi = { version = "0.6", default-features = false, features = ["derive"], optional = true }
 glam = { version = "0.21.3" }
 once_cell = { version = "1.13.0", optional = true }
+num-traits = "0.2.15"
